+++ my-react-portfolio/src/assets/components/Header/Header.css
@@ -5,9 +5,17 @@ nav {
     gap:1rem;
     width: 100%;
     font-family: Montserrat;
-    font-size: 32px;
+    font-size: 1.5rem;
     font-style: normal;
     line-height: normal;
     margin-bottom: 20px;
     color: #0B24F5;
+    
+}
+nav div{
+    transition: color 0.3s ease;
+}
+nav div:hover {
+    color: #FFA500; /* Example color change on hover */
+    cursor: pointer;
 }
