--- my-react-portfolio/src/assets/components/Tech.jsx
@@ -1,21 +1,0 @@
-export const Tech = () => {
-    return (
-        <div className="skills-wrapper">
-            <h3>WHAT I DO<br /><span>TECH</span></h3>
-            <p>HTML5</p>
-            <p>CSS3</p>
-            <p>JavaScript ES6</p>
-            <p>React</p>
-            <p>JSX</p>
-            <p>React Hooks</p>
-            <p>Web Accessibility</p>
-            <p>APIs</p>
-            <p>mob-programming</p>
-            <p>pair-programming</p>
-            <p>GitHub</p>
-            <p>Git</p>
-            <p>Node.js</p>
-            <p>Mongo DB</p>
-        </div>
-    )
-}
