+++ my-react-portfolio/src/assets/components/Tech/Tech.css
@@ -1,0 +1,32 @@
+.tech-wrapper{
+    display: flex;
+width: 1280px;
+height: 980px;
+padding: 0px 161px;
+justify-content: center;
+align-items: center;
+flex-shrink: 0;
+background-image: linear-gradient(to right, #FFA500, #0B24F5);
+transition: background-image 0.5s ease;
+}
+
+.tech-wrapper:hover{
+    background-image: linear-gradient(to right, #0B24F5, #FFA500 );
+}
+.tech-title{
+    color: #FFF;
+    font-family: Montserrat;
+    font-size: 80px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: normal;
+}
+
+.tech-details{
+    color: #FFF;
+    font-family: Hind;
+    font-size: 32px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: normal;
+}
